+++ app/Http/Controllers/MerchantsController.php
@@ -7,6 +7,7 @@
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
+
 namespace App\Http\Controllers;
 
 use App\Http\Requests;
@@ -22,14 +23,21 @@ use Illuminate\Database\Eloquent\ModelNotFoundException;
  */
 class MerchantsController extends Controller
 {
+    /** @var  \App\Basket\Synchronisation\MerchantSynchronisationService */
+    private $merchantSynchronisationService;
 
-	/**
-	 * Display a listing of the resource.
-	 *
-	 * @return Response
-	 */
-	public function index()
-	{
+    public function __construct()
+    {
+        $this->merchantSynchronisationService = \App::make('App\Basket\Synchronisation\MerchantSynchronisationService');
+    }
+
+    /**
+     * Display a listing of the resource.
+     *
+     * @return Response
+     */
+    public function index()
+    {
         $messages = $this->getMessages();
         $merchants = null;
 
@@ -46,25 +54,25 @@ class MerchantsController extends Controller
         }
 
         return View('merchants.index', ['merchants' => $merchants, 'messages' => $messages]);
-	}
+    }
 
-	/**
-	 * Show the form for creating a new resource.
-	 *
-	 * @return Response
-	 */
-	public function create()
-	{
+    /**
+     * Show the form for creating a new resource.
+     *
+     * @return Response
+     */
+    public function create()
+    {
         return view('merchants.create', ['messages' => $this->getMessages()]);
-	}
+    }
 
-	/**
-	 * Store a newly created resource in storage.
-	 *
-	 * @return Response
-	 */
-	public function store(Request $request)
-	{
+    /**
+     * Store a newly created resource in storage.
+     *
+     * @return Response
+     */
+    public function store(Request $request)
+    {
         $this->validate($request, [
             'name' => 'required',
             'token' => 'required',
@@ -74,7 +82,11 @@ class MerchantsController extends Controller
 
         try {
 
-            Merchant::create($request->all());
+            $merchant = Merchant::create($request->all());
+
+            $this->merchantSynchronisationService->synchroniseMerchant($merchant->id, true);
+
+            return redirect('merchants/' . $merchant->id)->with($message[0], $message[1]);
 
         } catch (ModelNotFoundException $e) {
 
@@ -83,16 +95,16 @@ class MerchantsController extends Controller
         }
 
         return redirect('merchants')->with($message[0], $message[1]);
-	}
+    }
 
-	/**
-	 * Display the specified resource.
-	 *
-	 * @param  int  $id
-	 * @return Response
-	 */
-	public function show($id)
-	{
+    /**
+     * Display the specified resource.
+     *
+     * @param  int  $id
+     * @return Response
+     */
+    public function show($id)
+    {
         $merchants = null;
         $messages = $this->getMessages();
 
@@ -109,16 +121,16 @@ class MerchantsController extends Controller
         }
 
         return view('merchants.show', ['merchants' => $merchants, 'messages' => $messages]);
-	}
+    }
 
-	/**
-	 * Show the form for editing the specified resource.
-	 *
-	 * @param  int  $id
-	 * @return Response
-	 */
-	public function edit($id)
-	{
+    /**
+     * Show the form for editing the specified resource.
+     *
+     * @param  int  $id
+     * @return Response
+     */
+    public function edit($id)
+    {
         $merchants = null;
         $messages = $this->getMessages();
 
@@ -135,16 +147,16 @@ class MerchantsController extends Controller
         }
 
         return view('merchants.edit', ['merchants' => $merchants, 'messages' => $messages]);
-	}
+    }
 
-	/**
-	 * Update the specified resource in storage.
-	 *
-	 * @param  int  $id
-	 * @return Response
-	 */
-	public function update($id, Request $request)
-	{
+    /**
+     * Update the specified resource in storage.
+     *
+     * @param  int  $id
+     * @return Response
+     */
+    public function update($id, Request $request)
+    {
         $message = ['success', 'Merchant details were successfully updated'];
 
         try {
@@ -161,17 +173,17 @@ class MerchantsController extends Controller
         }
 
         return redirect()->back()->with($message[0], $message[1]);
-	}
+    }
 
-	/**
-	 * Remove the specified resource from storage.
-	 *
-	 * @param  int  $id
-	 * @return Response
-	 */
-	public function destroy($id)
-	{
-        $message = ['success','Role was successfully deleted'];
+    /**
+     * Remove the specified resource from storage.
+     *
+     * @param  int  $id
+     * @return Response
+     */
+    public function destroy($id)
+    {
+        $message = ['success','Merchant was successfully deleted'];
         try {
 
             Merchant::destroy($id);
@@ -183,5 +195,21 @@ class MerchantsController extends Controller
         }
 
         return redirect('merchants')->with($message[0], $message[1]);
-	}
+    }
+
+    /**
+     * @param $id
+     * @return \Illuminate\Http\RedirectResponse
+     */
+    public function synchronise($id)
+    {
+        try {
+            $this->merchantSynchronisationService->synchroniseMerchant($id);
+            $message = ['success', 'Synchronisation complete successfully'];
+        } catch (\Exception $e) {
+            $this->logError('Error while trying to synchronise Merchant[' . $id . ']: ' . $e->getMessage());
+            $message = ['error', 'Synchronisation not complete successfully'];
+        }
+        return redirect('merchants/' . $id)->with($message[0], $message[1]);
+    }
 }