+++ resources/views/includes/form/select.blade.php
@@ -1,5 +1,3 @@
-<br>
-<hr class="hr-tight">
 @if($object->count())
     @foreach($object as $item)
         {{-- */$selects[]=$item->{$field};/* --}}
@@ -7,16 +5,18 @@
 @endif
 
 <select class="filter form-control" name="{{ $field }}" value="{!! Request::only($field)[$field] !!}">
-    <option></option>
+    <option value="">All</option>
     @if($object->count())
         @foreach(array_unique($selects) as $option)
             <option
                 @if(Request::only($field)[$field] == $option)
                     selected="selected"
                 @endif
-                    name="{{ $option }}">{{ $option }}</option>
+                    name="{{ $option }}">{{ ucwords($option) }}</option>
         @endforeach
     @else()
-        <option selected="selected" name="{!! Request::only($field)[$field] !!}">{!! Request::only($field)[$field] !!}</option>
+        <option selected="selected" name="{!! Request::only($field)[$field] !!}">
+            {!! Request::only($field)[$field] !!}
+        </option>
     @endif
 </select>