+++ CHANGELOG.md
@@ -6,6 +6,20 @@
 - Installation Entity
 - Installation Synchronisation Services
 - AbstractSynchronisationService
+- Added Installations Read & Update
+- Made Improvements to Breadcrumb
+- Refactored Installations and Merchants to use Basket Eloquent Models
+- Middleware to control if *User* can perform actions on a *Merchant*
+- *Merchant* synchronisation GUI
+- *Merchant* creation extended with synchronisation
+- Actions o a row helper for blade
+- Added Locations CRUD
+- Added Applications: Read views
+- Added pattern for drop downs on edit pages
+
+## Bug Fixes
+- Code Standards fixes
+- Fixed user migration for `merchant_id` `nullable()`
 
 # 0.2.0
 2015-07-17